@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:image_collage_widget/utils/CollageType.dart';
+import 'package:image_collage_widget/utils/collage_type.dart';
 import 'src/screens/collage_sample.dart';
 import 'src/tranistions/fade_route_transition.dart';
 
